Price as decimal

+++ src/Application/Products/Commands/UpdateProducts/UpdateProductCommand.cs
@@ -16,7 +16,7 @@ namespace DeliveryWebApp.Application.Products.Commands.UpdateProducts
         public int Id { get; set; }
         public int? Discount { get; set; }
         public int? Quantity { get; set; }
-        public double? Price { get; set; }
+        public decimal? Price { get; set; }
         public string Name { get; set; }
         public string Category { get; set; }
         public byte[] Image { get; set; }
@@ -52,7 +52,7 @@ namespace DeliveryWebApp.Application.Products.Commands.UpdateProducts
 
             if (request.Price != null)
             {
-                entity.Price = (double) request.Price;
+                entity.Price = (decimal) request.Price;
             }
 
             if (!string.IsNullOrEmpty(request.Category))
