+++ src/Domain/Entities/Product.cs
@@ -12,26 +12,26 @@ namespace DeliveryWebApp.Domain.Entities
     {
         public string Name { get; set; }
         public byte[] Image { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; } // FIXME: price value stored in database isn't correct
         public int Discount { get; set; }
         public string Category { get; set; }
         public int Quantity { get; set; }
         public int RestaurantId { get; set; }
         public Restaurant Restaurant { get; set; }
 
-        public double ApplyDiscount()
+        public decimal ApplyDiscount()
         {
             if (Discount != 0)
             {
-                return Price * ((100.00 - Discount) / 100.00);
+                return Price * ((100.00M - Discount) / 100.00M);
             }
 
             return Price;
         }
 
-        public static double TotalPrice(List<Product> list)
+        public static decimal TotalPrice(List<Product> list)
         {
-            const double tot = 0.00;
+            const decimal tot = 0.00M;
             return list.Count <= 0 ? tot : Math.Round(list.Sum(p => (p.ApplyDiscount() * p.Quantity)), 2);
         }
     }
